rename main component to courses-page component
