+++ angular-mentoring/src/app/app.component.html
@@ -1,8 +1,8 @@
 <section class="wrapp">
-  <app-header class="header"></app-header>
-  <app-breadcrumbs class="breadcrumbs"></app-breadcrumbs>
-  <app-main class="main"></app-main>
-  <app-footer class="footer"></app-footer>
+  <header class="header"></header>
+  <breadcrumbs class="breadcrumbs"></breadcrumbs>
+  <courses-page class="main"></courses-page>
+  <footer class="footer"></footer>
 </section>
 
 <!-- <router-outlet></router-outlet> -->
