+++ angular-mentoring/src/app/coursesPage/course.ts
