+++ angular-mentoring/src/app/coursesPage/courses-page.component.html
@@ -9,5 +9,5 @@
     <button mat-raised-button class="add__button add__button--color-blue box-shadow--none"><i class="material-icons">add</i>Add course</button>
 </div>
 <section class="main__list-courses">
-    <app-courses></app-courses>
+    <courses></courses>
 </section>
