+++ angular-mentoring/src/app/coursesPage/courses-page.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { MainComponent } from './main.component';
+import { MainComponent } from './courses-page.component';
 
 describe('MainComponent', () => {
   let component: MainComponent;
