+++ angular-mentoring/src/app/coursesPage/courses-page.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'courses-page',
+  templateUrl: './courses-page.component.html',
+  styleUrls: ['./courses-page.component.css']
+})
+export class CoursesPageComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
