+++ angular-mentoring/src/app/coursesPage/i-course.ts
