--- angular-mentoring/src/app/main/main.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-main',
-  templateUrl: './main.component.html',
-  styleUrls: ['./main.component.css']
-})
-export class MainComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
